# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_oojoseph67_ecommerce_internal_dto.CategoryResponse:
    properties:
      description:
        example: Electronic devices and gadgets
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Electronics
        type: string
    type: object
...
  github_com_oojoseph67_ecommerce_internal_dto.CreateProductRequest:
    properties:
      category_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      description:
        example: Premium wireless headphones with noise cancellation
        type: string
      name:
        example: Wireless Headphones
        type: string
      price:
        example: 99.99
        type: number
      sku:
        example: WH-1000XM4
        type: string
      stock:
        example: 100
        minimum: 0
        type: integer
    required:
    - category_id
    - description
    - name
    - price
    - sku
    - stock
    type: object
...
  github_com_oojoseph67_ecommerce_internal_dto.ProductImageResponse:
    properties:
      alt_text:
        example: Product image
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_primary:
        example: true
        type: boolean
      url:
        example: https://storage.example.com/products/image.jpg
        type: string
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.CategoryResponse'
      category_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      description:
        example: Premium wireless headphones with noise cancellation
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      images:
        items:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.ProductImageResponse'
        type: array
      is_active:
        example: true
        type: boolean
      meta:
        $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.ResponseMeta'
      name:
        example: Wireless Headphones
        type: string
      price:
        example: 99.99
        type: number
      sku:
        example: WH-1000XM4
        type: string
      stock:
        example: 100
        type: integer
    type: object
...
  github_com_oojoseph67_ecommerce_internal_dto.ResponseMeta:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
    type: object
...
  github_com_oojoseph67_ecommerce_internal_dto.UpdateProductRequest:
    properties:
      category_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      description:
        example: Premium wireless headphones with noise cancellation
        type: string
      is_active:
        example: true
        type: boolean
      name:
        example: Wireless Headphones
        type: string
      price:
        example: 149.99
        type: number
      stock:
        example: 50
        minimum: 0
        type: integer
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.UpdateProfileRequest:
    properties:
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
      phone:
        example: "+1234567890"
        type: string
    type: object
  github_com_oojoseph67_ecommerce_internal_dto.UserResponse:
    properties:
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_active:
        example: true
        type: boolean
      last_name:
        example: Doe
        type: string
      phone:
        example: "+1234567890"
        type: string
      role:
        example: user
        type: string
    type: object
  github_com_oojoseph67_ecommerce_internal_utils_responses.PaginatedResponse:
    properties:
      meta:
        $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.PaginationMeta'
      response:
        $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
    type: object
  github_com_oojoseph67_ecommerce_internal_utils_responses.PaginationMeta:
    properties:
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  github_com_oojoseph67_ecommerce_internal_utils_responses.Response:
    properties:
      data: {}
      error:
        example: invalid request
        type: string
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:1234
info:
  contact:
    email: oojoseph67@gmail.com
    name: Joseph
    url: https://josephbuilds.xyz
  description: A modern e-commerce API built with Go, Gin, and GORM
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: E-Commerce API
  version: "1"
paths:
...
  /products/:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of products with optional category filter
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Category ID filter
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product (admin only)
      parameters:
      - description: Product data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Create a product
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Retrieve a single product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.ProductResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Get a product
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update an existing product (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - Products
  /products/{id}/upload:
    delete:
      consumes:
      - application/json
      description: Delete a product image (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Delete product image
      tags:
      - Products
    post:
      consumes:
      - multipart/form-data
      description: Upload one or more images for a product (admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Image files to upload
        in: formData
        name: productImage
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Upload product images
      tags:
      - Products
  /user/me:
    get:
      consumes:
      - application/json
      description: Retrieve the authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - User
  /user/update:
    patch:
      consumes:
      - application/json
      description: Update the authenticated user's profile information
      parameters:
      - description: Profile update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_oojoseph67_ecommerce_internal_utils_responses.Response'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - User
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"